[Refactor] rearrange files and remove redundant function
