+++ socket/services/service.js
@@ -3,6 +3,22 @@ const { SocketException } = require('../../server/services/exceptions/socketExce
 const { jwtAuthenticate } = require('../../server/services/auth');
 const Cache = require('../../utils/cache');
 
+const CLIENT_CATEGORY = {
+  visitor: 0,
+  otherMember: 1,
+  self: 2,
+};
+
+const authorization = (userID, userPayload) => {
+  if (!userPayload) {
+    return CLIENT_CATEGORY.visitor;
+  }
+  if (userID !== userPayload) {
+    return CLIENT_CATEGORY.otherMember;
+  }
+  return CLIENT_CATEGORY.self;
+};
+
 function wrapAsync(cb) {
   return async function (emitObject) {
     const socket = this;
@@ -44,5 +60,10 @@ function checkCacheReady() {
 }
 
 module.exports = {
-  wrapAsync, wrapMiddleware, socketAuth, checkCacheReady,
+  wrapAsync,
+  wrapMiddleware,
+  socketAuth,
+  checkCacheReady,
+  CLIENT_CATEGORY,
+  authorization,
 };
